added main.css
