+++ website/components/tab/tab.module.css
@@ -1,0 +1,3 @@
+@tailwind components;
+
+
