+++ website/pages/_app.js
@@ -1,4 +1,5 @@
 import 'tailwindcss/tailwind.css'
+import '../main.css'
 
 function MyApp({ Component, pageProps }) {
   return (
